update XFSTesting and Xfstests tool
