+++ lisa/tools/__init__.py
@@ -129,7 +129,6 @@ from .who import Who
 from .whoami import Whoami
 from .windows_feature import WindowsFeatureManagement
 from .wsl import Wsl
-from .xfstests import Xfstests
 
 __all__ = [
     "AptAddRepository",
@@ -270,5 +269,4 @@ __all__ = [
     "Whoami",
     "WindowsFeatureManagement",
     "Wsl",
-    "Xfstests",
 ]
